fix tests [3]
