+++ library/core/Curl/Exception.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Swoole\Curl;
+
+use Swoole;
+
+class Exception extends Swoole\Exception
+{
+
+}
