+++ php_swoole_library.h
@@ -62,10 +62,24 @@ static const char* swoole_library_source_std_exec =
     "    return null;\n"
     "}\n";
 
-static const char* swoole_library_source_ext_curl =
+static const char* swoole_library_source_core_curl_exception =
     "\n"
     "\n"
-    "abstract class swoole_http_status_code\n"
+    "namespace Swoole\\Curl;\n"
+    "\n"
+    "use Swoole;\n"
+    "\n"
+    "class Exception extends Swoole\\Exception\n"
+    "{\n"
+    "\n"
+    "}\n";
+
+static const char* swoole_library_source_core_http_status_code =
+    "\n"
+    "\n"
+    "namespace Swoole\\Http;\n"
+    "\n"
+    "abstract class StatusCode\n"
     "{\n"
     "    const CONTINUE = 100;\n"
     "    const SWITCHING_PROTOCOLS = 101;\n"
@@ -202,15 +216,28 @@ static const char* swoole_library_source_ext_curl =
     "    {\n"
     "        return static::$reasonPhrases[$value] ?? 'Unknown';\n"
     "    }\n"
-    "}\n"
+    "}\n";
+
+static const char* swoole_library_source_core_curl_handler =
     "\n"
-    "class swoole_curl_handler\n"
+    "\n"
+    "namespace Swoole\\Curl;\n"
+    "\n"
+    "\n"
+    "use Swoole;\n"
+    "use Swoole\\Coroutine\\Http\\Client;\n"
+    "use CURLFile;\n"
+    "use Iterator;\n"
+    "\n"
+    "class Handler\n"
     "{\n"
     "    private const ERRORS = [\n"
     "        CURLE_URL_MALFORMAT => 'No URL set!',\n"
     "    ];\n"
     "\n"
-    "    /** @var Swoole\\Coroutine\\Http\\Client */\n"
+    "    /**\n"
+    "     * @var Client\n"
+    "     */\n"
     "    private $client;\n"
     "    private $info = [\n"
     "        'url' => '',\n"
@@ -295,7 +322,7 @@ static const char* swoole_library_source_ext_curl =
     "            $port = intval($info['port']);\n"
     "        }\n"
     "        $this->urlInfo = $info;\n"
-    "        $this->client = new Swoole\\Coroutine\\Http\\Client($info['host'], $port, $ssl);\n"
+    "        $this->client = new Client($info['host'], $port, $ssl);\n"
     "    }\n"
     "\n"
     "    public function execute()\n"
@@ -319,7 +346,7 @@ static const char* swoole_library_source_ext_curl =
     "                $proto = swoole_array_default_value($this->urlInfo, 'scheme');\n"
     "                if ($proto != 'http' and $proto != 'https') {\n"
     "                    $this->setError(CURLE_UNSUPPORTED_PROTOCOL, \"Protocol \\\"{$proto}\\\" not supported or disabled in libcurl\");\n"
-    "                    return;\n"
+    "                    return false;\n"
     "                }\n"
     "                $ssl = $proto === 'https';\n"
     "                if (empty($this->urlInfo['port'])) {\n"
@@ -327,7 +354,7 @@ static const char* swoole_library_source_ext_curl =
     "                } else {\n"
     "                    $port = intval($this->urlInfo['port']);\n"
     "                }\n"
-    "                $client = new Swoole\\Coroutine\\Http\\Client($this->urlInfo['host'], $port, $ssl);\n"
+    "                $client = new Client($this->urlInfo['host'], $port, $ssl);\n"
     "            }\n"
     "            /**\n"
     "             * Http Proxy\n"
@@ -335,7 +362,7 @@ static const char* swoole_library_source_ext_curl =
     "            if ($this->proxy) {\n"
     "                list($proxy_host, $proxy_port) = explode(':', $this->proxy);\n"
     "                if (!filter_var($proxy_host, FILTER_VALIDATE_IP)) {\n"
-    "                    $ip = \\Swoole\\Coroutine::gethostbyname($proxy_host);\n"
+    "                    $ip = Swoole\\Coroutine::gethostbyname($proxy_host);\n"
     "                    if (!$ip) {\n"
     "                        $this->setError(CURLE_COULDNT_RESOLVE_PROXY, 'Could not resolve proxy: ' . $proxy_host);\n"
     "                        return false;\n"
@@ -384,7 +411,7 @@ static const char* swoole_library_source_ext_curl =
     "            $executeResult = $client->execute($this->getUrl());\n"
     "            if (!$executeResult) {\n"
     "                $errCode = $client->errCode;\n"
-    "                if ($errCode == 1 and $client->errMsg == 'Unknown host') {\n"
+    "                if ($errCode == SWOOLE_ERROR_DNSLOOKUP_RESOLVE_FAILED or $errCode == SWOOLE_ERROR_DNSLOOKUP_RESOLVE_TIMEOUT) {\n"
     "                    $this->setError(CURLE_COULDNT_RESOLVE_HOST, 'Could not resolve host: ' . $client->host);\n"
     "                }\n"
     "                $this->info['total_time'] = microtime(true) - $timeBegin;\n"
@@ -427,26 +454,26 @@ static const char* swoole_library_source_ext_curl =
     "            $this->info['redirect_time'] = microtime(true) - $redirectBeginTime;\n"
     "        }\n"
     "\n"
+    "        $headerContent = '';\n"
     "        if ($client->headers) {\n"
     "            $cb = $this->headerFunction;\n"
-    "            $headerContent = '';\n"
     "            if ($client->statusCode > 0) {\n"
-    "                $row = 'HTTP/1.1 ' . $client->statusCode . ' ' . swoole_http_status_code::getReasonPhrase($client->statusCode) . \"\\r\\n\";\n"
-    "                if($cb) {\n"
+    "                $row = 'HTTP/1.1 ' . $client->statusCode . ' ' . Swoole\\Http\\StatusCode::getReasonPhrase($client->statusCode) . \"\\r\\n\";\n"
+    "                if ($cb) {\n"
     "                    $cb($this, $row);\n"
     "                }\n"
     "                $headerContent .= $row;\n"
     "            }\n"
     "            foreach ($client->headers as $k => $v) {\n"
     "                $row = \"$k: $v\\r\\n\";\n"
-    "                if($cb) {\n"
+    "                if ($cb) {\n"
     "                    $cb($this, $row);\n"
     "                }\n"
     "                $headerContent .= $row;\n"
     "            }\n"
     "            $headerContent .= \"\\r\\n\";\n"
     "            $this->info['header_size'] = strlen($headerContent);\n"
-    "            if($cb) {\n"
+    "            if ($cb) {\n"
     "                $cb($this, '');\n"
     "            }\n"
     "        } else {\n"
@@ -458,7 +485,7 @@ static const char* swoole_library_source_ext_curl =
     "            $cb($this, $this->outputStream, strlen($client->body));\n"
     "        }\n"
     "\n"
-    "        if($this->withHeader) {\n"
+    "        if ($this->withHeader) {\n"
     "            $transfer = $headerContent . $client->body;\n"
     "        } else {\n"
     "            $transfer = $client->body;\n"
@@ -507,7 +534,7 @@ static const char* swoole_library_source_ext_curl =
     "     * @param int $opt\n"
     "     * @param $value\n"
     "     * @return bool\n"
-    "     * @throws swoole_curl_exception\n"
+    "     * @throws Swoole\\Curl\\Exception\n"
     "     */\n"
     "    public function setOption(int $opt, $value): bool\n"
     "    {\n"
@@ -576,7 +603,7 @@ static const char* swoole_library_source_ext_curl =
     "                break;\n"
     "            case CURLOPT_PROTOCOLS:\n"
     "                if ($value > 3) {\n"
-    "                    throw new swoole_curl_exception(\"option[{$opt}={$value}] not supported\");\n"
+    "                    throw new Swoole\\Curl\\Exception(\"option[{$opt}={$value}] not supported\");\n"
     "                }\n"
     "                break;\n"
     "            case CURLOPT_HTTP_VERSION:\n"
@@ -635,7 +662,7 @@ static const char* swoole_library_source_ext_curl =
     "                $this->maxRedirs = $value;\n"
     "                break;\n"
     "            default:\n"
-    "                throw new swoole_curl_exception(\"option[{$opt}] not supported\");\n"
+    "                throw new Swoole\\Curl\\Exception(\"option[{$opt}] not supported\");\n"
     "        }\n"
     "        return true;\n"
     "    }\n"
@@ -651,14 +678,14 @@ static const char* swoole_library_source_ext_curl =
     "\n"
     "    private function unparseUrl(array $parsedUrl): string\n"
     "    {\n"
-    "        $scheme   = ($parsedUrl['scheme'] ?? 'http') . '://';\n"
-    "        $host     = $parsedUrl['host'] ?? '';\n"
-    "        $port     = isset($parsedUrl['port']) ? ':' . $parsedUrl['port'] : '';\n"
-    "        $user     = $parsedUrl['user'] ?? '';\n"
-    "        $pass     = isset($parsedUrl['pass']) ? ':' . $parsedUrl['pass']  : '';\n"
-    "        $pass     = ($user or $pass) ? \"$pass@\" : '';\n"
-    "        $path     = $parsedUrl['path'] ?? '';\n"
-    "        $query    = (isset($parsedUrl['query']) and '' !== $parsedUrl['query']) ? '?' . $parsedUrl['query'] : '';\n"
+    "        $scheme = ($parsedUrl['scheme'] ?? 'http') . '://';\n"
+    "        $host = $parsedUrl['host'] ?? '';\n"
+    "        $port = isset($parsedUrl['port']) ? ':' . $parsedUrl['port'] : '';\n"
+    "        $user = $parsedUrl['user'] ?? '';\n"
+    "        $pass = isset($parsedUrl['pass']) ? ':' . $parsedUrl['pass'] : '';\n"
+    "        $pass = ($user or $pass) ? \"$pass@\" : '';\n"
+    "        $path = $parsedUrl['path'] ?? '';\n"
+    "        $query = (isset($parsedUrl['query']) and '' !== $parsedUrl['query']) ? '?' . $parsedUrl['query'] : '';\n"
     "        $fragment = isset($parsedUrl['fragment']) ? '#' . $parsedUrl['fragment'] : '';\n"
     "        return $scheme . $user . $pass . $host . $port . $path . $query . $fragment;\n"
     "    }\n"
@@ -694,37 +721,35 @@ static const char* swoole_library_source_ext_curl =
     "        }\n"
     "        return $redirectUri;\n"
     "    }\n"
-    "}\n"
+    "}\n";
+
+static const char* swoole_library_source_ext_curl =
     "\n"
-    "class swoole_curl_exception extends swoole_exception\n"
+    "\n"
+    "function swoole_curl_init($url = null): Swoole\\Curl\\Handler\n"
     "{\n"
-    "\n"
-    "}\n"
-    "\n"
-    "function swoole_curl_init($url = null): swoole_curl_handler\n"
-    "{\n"
-    "    return new swoole_curl_handler($url);\n"
+    "    return new Swoole\\Curl\\Handler($url);\n"
     "}\n"
     "\n"
     "/**\n"
-    " * @param swoole_curl_handler $obj\n"
+    " * @param Swoole\\Curl\\Handler $obj\n"
     " * @param $opt\n"
     " * @param $value\n"
     " * @return bool\n"
-    " * @throws swoole_curl_exception\n"
+    " * @throws Swoole\\Curl\\Exception\n"
     " */\n"
-    "function swoole_curl_setopt(swoole_curl_handler $obj, $opt, $value): bool\n"
+    "function swoole_curl_setopt(Swoole\\Curl\\Handler $obj, $opt, $value): bool\n"
     "{\n"
     "    return $obj->setOption($opt, $value);\n"
     "}\n"
     "\n"
     "/**\n"
-    " * @param swoole_curl_handler $obj\n"
+    " * @param Swoole\\Curl\\Handler $obj\n"
     " * @param $array\n"
     " * @return bool\n"
-    " * @throws swoole_curl_exception\n"
+    " * @throws Swoole\\Curl\\Exception\n"
     " */\n"
-    "function swoole_curl_setopt_array(swoole_curl_handler $obj, $array): bool\n"
+    "function swoole_curl_setopt_array(Swoole\\Curl\\Handler $obj, $array): bool\n"
     "{\n"
     "    foreach ($array as $k => $v) {\n"
     "        if ($obj->setOption($k, $v) === false) {\n"
@@ -734,32 +759,32 @@ static const char* swoole_library_source_ext_curl =
     "    return true;\n"
     "}\n"
     "\n"
-    "function swoole_curl_exec(swoole_curl_handler $obj)\n"
+    "function swoole_curl_exec(Swoole\\Curl\\Handler $obj)\n"
     "{\n"
     "    return $obj->execute();\n"
     "}\n"
     "\n"
-    "function swoole_curl_close(swoole_curl_handler $obj): void\n"
+    "function swoole_curl_close(Swoole\\Curl\\Handler $obj): void\n"
     "{\n"
     "    $obj->close();\n"
     "}\n"
     "\n"
-    "function swoole_curl_error(swoole_curl_handler $obj): string\n"
+    "function swoole_curl_error(Swoole\\Curl\\Handler $obj): string\n"
     "{\n"
     "    return $obj->errMsg;\n"
     "}\n"
     "\n"
-    "function swoole_curl_errno(swoole_curl_handler $obj): int\n"
+    "function swoole_curl_errno(Swoole\\Curl\\Handler $obj): int\n"
     "{\n"
     "    return $obj->errCode;\n"
     "}\n"
     "\n"
-    "function swoole_curl_reset(swoole_curl_handler $obj): void\n"
+    "function swoole_curl_reset(Swoole\\Curl\\Handler $obj): void\n"
     "{\n"
     "    $obj->reset();\n"
     "}\n"
     "\n"
-    "function swoole_curl_getinfo(swoole_curl_handler $obj, int $opt = 0)\n"
+    "function swoole_curl_getinfo(Swoole\\Curl\\Handler $obj, int $opt = 0)\n"
     "{\n"
     "    $info = $obj->getInfo();\n"
     "    if ($opt) {\n"
@@ -2006,18 +2031,6 @@ static const char* swoole_library_source_functions =
     "    {\n"
     "        return new Swoole\\ArrayObject($array);\n"
     "    }\n"
-    "\n"
-    "    /**\n"
-    "     * @return Swoole\\Coroutine\\Scheduler\n"
-    "     */\n"
-    "    function scheduler()\n"
-    "    {\n"
-    "        static $scheduler = null;\n"
-    "        if (!$scheduler) {\n"
-    "            $scheduler = new Swoole\\Coroutine\\Scheduler();\n"
-    "        }\n"
-    "        return $scheduler;\n"
-    "    }\n"
     "}\n"
     "\n"
     "/**\n"
@@ -2086,6 +2099,9 @@ static void php_swoole_load_library()
     zend::eval(swoole_library_source_constants, "@swoole-src/library/constants.php");
     zend::eval(swoole_library_source_std_array, "@swoole-src/library/std/array.php");
     zend::eval(swoole_library_source_std_exec, "@swoole-src/library/std/exec.php");
+    zend::eval(swoole_library_source_core_curl_exception, "@swoole-src/library/core/Curl/Exception.php");
+    zend::eval(swoole_library_source_core_http_status_code, "@swoole-src/library/core/Http/StatusCode.php");
+    zend::eval(swoole_library_source_core_curl_handler, "@swoole-src/library/core/Curl/Handler.php");
     zend::eval(swoole_library_source_ext_curl, "@swoole-src/library/ext/curl.php");
     zend::eval(swoole_library_source_core_coroutine_wait_group, "@swoole-src/library/core/Coroutine/WaitGroup.php");
     zend::eval(swoole_library_source_core_coroutine_object_pool, "@swoole-src/library/core/Coroutine/ObjectPool.php");
